Lumap V1.1
This is still the original version, which extracts the source code from the PDF and stores it in five documents, making it easier to download and run. And for v errors in version 1.0 were fixed, including logic errors in the chemicals management module, code specification and maintenance of the risk monitoring module, and functional loss of the front page. Not much change in overall functionality